Use pause_before option in remote shell provisioner
The pause_before option is a newer option that makes managing the
problem of reboots in scripts easier, with fewer hangups that before.
Previously, the recommendation was to put a reboot in a script immediately
followed by a sleep. What would happen, far too frequently, is the
reboot would kick off successfully but not the sleep. Vagrant would
start executing the second script, which would then hangup during
the reboot and the build would never complete.