Reorganize scripts directory
The previous commit attempted to use one updatePackages.sh script
for both CentOS and Ubuntu VMs. It works, but opens the door to
code duplication throughout scripts/.

For example, anywhere a package install or update is needed, the
same check for yum vs. apt-get would have to be performed. Similar
situations are possible, and though there aren't many scripts now
there very well could be in the future and it's best to eliminate
the problem now before it expands.

The possibility for a scripts/generic/ directory exists in the future.
Scripts that are exactly the same between distributions or scripts
that could be refactored to be the same could go there to further
reduce code duplication.