Make builds headless
A GUI can be a great asset for troubleshooting a failing build but
isn't necessary for typical builds. From experience, users can't
seem to resist messing with the VM when a GUI is present and that
can cause the build to fail.

The GUI can be re-enabled by setting the headless option to false.