Add working template with amazon-import post-processor
This commit adds a separate template for provisioning a base CentOS 7
AMI with Amazon Web Services. The template will work, provided that
the variables are set correctly. In particular, make sure the given
S3 bucket exists and that you have the vmimport IAM role (see
http://docs.aws.amazon.com/vm-import/latest/userguide/import-vm-image.html
for more).

It's important to note that the AMI that is created is what you provide,
meaning that no user creation and key setup is done like with other AMIs
in the AWS marketplace. Cloud-Init could/should be added to do such
provisioning and more upon creation of an EC2 instance.