Create centos-7 OVA base job
Both the amazon-import and vagrant-virtualbox jobs require an OS
to be installed and Yum updates to be applied. Rather than do that each
time for each build we can export an OVA of an installed OS with Yum
updates applied and continue from that point. It'll save time and also
guarantees that all centos-7 builders start from the same patch level.