Improve amazon-import security
Cloud-init will block SSH as root, so while it's not strictly necessary
to set a random root password it's better than the default 'toor'
password.

SELinux should be in enforcing mode, especially for cloud-hosted
servers. Currently, the kickstart process starts SELinux in permissive
mode, which may be a legacy artifact that's no longer necessary. It will
be reviewed separately.