Set SELinux to enforcing
By default, kickstart will set SELinux to enforcing mode. It's
beneficial to develop on Vagrant boxes with SELinux in enforcing mode to
expose any security-relevant changes that need to be made prior to going
to QA or PROD. This change also ensures that the
centos-7/amazon-import.json template builds an AMI that is security
ready for deployment to a cloud-hosted environment.